app/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-telegram-bot==22.1
cryptography==42.0.8
yfinance==0.2.48
APScheduler==3.10.4
aiohttp==3.9.1
requests==2.31.0
python-dotenv==1.0.0